+++ modules/server/src/Application.ts
@@ -1,9 +1,7 @@
 import { InternalServer } from './internal';
 import { Dependency } from './dependency';
 import { ILogger, ConsoleLogger } from './loggers';
-import { IController } from './controller';
-import { IInjector } from './Injector';
-import { IServiceHandler, IServerConfig } from './interfaces';
+import { IServiceHandler, IServerConfig, IController, IInjector, IApplication, IDependency } from './interfaces';
 import * as isPromise from 'is-promise';
 import * as http2 from 'http2';
 import * as http from 'http';
@@ -15,7 +13,7 @@ import * as glob from 'glob';
  * based on the routes and attributes of the controllers.  The main purpose in this is so that
  * webpack and normal registry of modules that house the controllers can be found by this class.
  */
-export class Application {
+export class Application implements IApplication {
     /**
      * The directory that this server is currently running in and where it should look for the controllers.
      */
@@ -57,12 +55,13 @@ export class Application {
      * @param logger The logger that should be used will default to a console logger.
      * @param controllerSuffix The controller suffix that should be used if not using the webpack require api.
      * @param injectorSuffix The injector suffix that should be used if not using the webpack require api.
+     * @param startUpSuffix The startup suffix to get the startup function from applications that do not use webpack require api.
      */
-    constructor(config: string | IServerConfig, serviceHandler?: IServiceHandler, logger?: ILogger, public controllerSuffix: string = 'controller', public injectorSuffix: string = 'injector') {
+    constructor(config: string | IServerConfig, serviceHandler?: IServiceHandler, logger?: ILogger, public controllerSuffix: string = 'controller', public injectorSuffix: string = 'injector', public startUpSuffix: string = 'startup') {
         config = typeof config === 'string' ? { cwd: config } : config;
 
         this._bootStrapped = false;
-        this._config = <IServerConfig>config;
+        this._config = config as IServerConfig;
         this._logger = logger ? logger : new ConsoleLogger();
         this._dependency = new Dependency();
         this._serviceHandler = serviceHandler;
@@ -103,6 +102,19 @@ export class Application {
         let items = await this.findItems<IInjector>(context, dirname || this._config.cwd, this.injectorSuffix);
         for (let i = 0; i < items.length; ++i) {
             this._server.addInjectors(items[i]);
+        }
+    }
+
+    /**
+     * Method is meant to call start up functions to allow for modules to add in dependencies, injectors and middleware to the server.
+     * 
+     * @param context The context in which to find the start up files.
+     * @param dirname The directory name that should be used instead o the current directory the server is running in.
+     */
+    public async startup(context: string | __WebpackModuleApi.RequireContext, dirname?: string): Promise<void> {
+        let items = await this.findItems<(app: IApplication) => void>(context, dirname || this._config.cwd, this.startUpSuffix);
+        for (let i = 0; i < items.length; ++i) {
+            items[i](this);
         }
     }
 
@@ -184,6 +196,52 @@ export class Application {
     public middleware(callback: (req: http2.Http2ServerRequest, res: http2.Http2ServerResponse, next: (err?: any) => void) => void);
     public middleware(callback: (req: any, res: any, next: (err?: any) => void) => void) {
         this._server.middleware(callback);
+    }
+
+    /**
+     * Method is meant to add instances of the object instead of the class type.
+     * 
+     * @param injectable The instance of the object that needs to be included as an injectable.
+     */
+    public addSingle(injectable: any): void {
+        this._dependency.addSingle(injectable);
+    }
+
+    /**
+     * Method is meant to include a class type into the injectables but first resolve the class before
+     * creating an instance of the class.
+     * 
+     * @param dependency The type of the class that needs to be included into the injectables.
+     */
+    public addDependency(dependency: IDependency): void {
+        this._dependency.addDependency(dependency);
+    }
+
+    /**
+     * Method is meant to include dependencies that should be used during a scope.
+     * 
+     * @param dependency The dependency that we need to create when a scope is started.
+     */
+    public addScoped(dependency: IDependency): void {
+        this._dependency.addScoped(dependency);
+    }
+
+    /**
+     * Method is meant to handle the additions of the controllers to the server.
+     * 
+     * @param controllers The controllers that need to be added to the server for use.
+     */
+    public addControllers(...controllers: IController[]): void {
+        this._server.addControllers(...controllers);
+    }
+
+    /**
+     * Method is meant to handle the additions of the injectors to the server.
+     * 
+     * @param injectors The injectors that need the be added to the server for use.
+     */
+    public addInjectors(...injectors: IInjector[]): void {
+        this._server.addInjectors(...injectors);
     }
 
     /**