+++ modules/server/src/interfaces/IController.ts
@@ -1,0 +1,12 @@
+import { IRoute } from './IRoute';
+import { IInternalRoute } from './IInternalRoute';
+import { Controller } from '../controller';
+
+/**
+ * Helper interface that is used internaly so that typescript builds properly.
+ */
+export interface IController {
+  new (...args: any[]): Controller;
+  __routes__?: IRoute[];
+  generateRoutes(): IInternalRoute[];
+}