+++ modules/server/src/interfaces/IDependency.ts
@@ -1,0 +1,6 @@
+/**
+ * Dependency that requires the object to be a class or a newable function.
+ */
+export interface IDependency {
+  new (...args: any[]): any;
+}