+++ modules/server/src/interfaces/IInjector.ts
@@ -1,0 +1,12 @@
+import { IRoute } from './IRoute';
+import { IInternalInjectedRoute } from './IInternalInjectedRoute';
+import { Injector } from '../Injector';
+
+/**
+ * Helper interface that is used internaly so that typescript builds properly.
+ */
+export interface IInjector {
+  new (...args: any[]): Injector;
+  __routes__?: IRoute[];
+  generateRoutes(): IInternalInjectedRoute[];
+}